remoteproc: qcom_wcnss: Add pm domain specific regulator descriptor
This patch adds a pm domain specific regulator descriptor.
When using regulator corners on pronto-v2 hardware we need to map

- vddmx
- vddcx

as RPMD corner regulators. In contrast

- vddpx

is controlled via a PMIC as is the case on MSM8916 and MSM8997.

This patch adds a binding called "qcom,pronto-v2-pd-pil" and a probe time
table to go with that binding, describing pronto-v2 hardware with a
different regulator API.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>
Signed-off-by: Pavel Dubrova <pashadubrova@gmail.com>